% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatVictor.R
\name{formatVictor}
\alias{formatVictor}
\title{Read florescence data from PerkinElmer Victor and format it}
\usage{
formatVictor(path)
}
\arguments{
\item{path}{path to data file}
}
\value{
list of metadata and measurement table
}
\description{
Read florescence data from PerkinElmer Victor and format it
}
\examples{
formatVictor("~/Downloads/BCA.txt")
}
